Refactor tests into macro
